move contexts to folders
